Renaming run.py to runner.py, revert horovodrun rename
- horovod/run/run.py and horovod/spark/run.py renamed
- revert horovodrun renaming in strings, will do in separate PR
