+++ horovod/run/__init__.py
@@ -12,3 +12,5 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 # =============================================================================
+
+from .runner import run
