+++ horovod/run/js_run.py
@@ -56,7 +56,7 @@ def js_run(settings, nics, env, command, stdout=None, stderr=None, run_func=safe
 
     if not is_jsrun_installed():
         raise Exception(
-            'horovod does not find the jsrun command.\n\n'
+            'horovodrun convenience script does not find the jsrun command.\n\n'
             'Please, make sure you are running on a cluster with jsrun installed or '
             'use one of the other launchers.')
