+++ horovod/spark/runner.py
@@ -23,7 +23,7 @@ from six.moves import queue
 from horovod.spark.task import task_service
 from horovod.spark.gloo_run import gloo_run
 from horovod.spark.mpi_run import mpi_run
-from horovod.run.run import run_controller
+from horovod.run.runner import run_controller
 from horovod.run.common.util import timeout, host_hash, secret
 from horovod.run.common.util import settings as hvd_settings
 from horovod.spark.driver import driver_service, job_id
